requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
# AI/CV dependencies
torch
torchvision
rembg[cpu]
scikit-learn
numpy
Pillow
opencv-python-headless
